@@ -35,7 +35,7 @@ export class ProjectMonitoringComponent implements OnInit, OnDestroy, AfterViewI
   private subs = new SubSink();
   private timer: NodeJS.Timer;
   private sourceInterval: NodeJS.Timer;
-  @ViewChild(MatSort) sort: MatSort;
+  @ViewChild(MatSort, {static: false}) sort: MatSort;
   private columnToSort: string;
   private sortDirection: SortDirection;
 
